Grammar: Support ASP server-side comments
Originally reported as an issue with GitHub's linguist, I found
that the problem was actually with the HTML-ASP textmate bundle
that GitHub's Linguist uses for syntax highlighting.

I've added a pattern which detects the <%-- --%> syntax used for
ASP server-side comments (as used in ASP and ASP.NET).

This is linked to and fixes github/linguist#895.